+++ client/src/main.jsx
@@ -23,7 +23,6 @@ const router = createBrowserRouter([
         element: <Signin />,
       },
       {
-        
         path: "/forgotpassword",
         element: <Forget />,
       },
